Day 12 solution
This was the first day I struggled; I had to get a hint to use
Bredth First Search, and then realized that with a BFS, it's
simple because the first time you see each point gives you the
shortest distance to that point. My Depth First Search worked, but
even with crappy optimisations probably would have taken until the
heat death of the universe to find the solution.